Fixed confusion about enemies in fights.

+++ Source/Fight_MovementManager.h
@@ -57,7 +57,7 @@ public:
 
 
 private:
-    int distance_metric = 0;
+    double distance_metric = 0;
     Position stutter_vector_ = Positions::Origin;
     Position attune_vector_ = Positions::Origin;
     Position cohesion_vector_ = Positions::Origin;
